feat: add relation icon hover clicked event
