--- src/components/implements-icon/icon.type.ts
@@ -1,9 +1,0 @@
-// 식별 1:1, 1:n, m,n, 비식별 1:1, 1:n, m,n
-export type MappingType =
-  | 'IdentifyOneToOne'
-  | 'IdentifyOneToMany'
-  | 'IdentifyManyToMany'
-  | 'NonIdentifyOneToOne'
-  | 'NonIdentifyOneToMany'
-  | 'NonIdentifyManyToMany'
-  | 'none';
