+++ src/components/implements-icon/index.ts
@@ -1,5 +1,4 @@
 export * from './RelationshipIcons';
-export * from './icon.type';
 export * from './ArrowIcon';
 export * from './MemoIcon';
 export * from './PointerIcon';
