+++ src/features/draw-tools/draw-tools.slice.ts
@@ -3,16 +3,26 @@ import { create } from 'zustand';
 type CursorType = 'arrow' | 'pointer';
 type EntityType = 'table' | 'memo' | 'none';
 
+// 식별 1:1, 1:n, m,n, 비식별 1:1, 1:n, m,n
+export type MappingType =
+  | 'IdentifyOneToOne'
+  | 'IdentifyOneToMany'
+  | 'IdentifyManyToMany'
+  | 'NonIdentifyOneToOne'
+  | 'NonIdentifyOneToMany'
+  | 'NonIdentifyManyToMany'
+  | 'none';
+
 export interface ToolsChoose {
   cursor: CursorType;
   entity: EntityType;
-  mapping: number;
+  mapping: MappingType;
 }
 
 export interface ToolsChooseAction {
   setCursor: (cursor: CursorType) => void;
   setEntity: (entity: EntityType) => void;
-  setMapping: (mapping: number) => void;
+  setMapping: (mapping: MappingType) => void;
 }
 
 export type DrawToolsSlice = ToolsChoose & ToolsChooseAction;
@@ -20,7 +30,7 @@ export type DrawToolsSlice = ToolsChoose & ToolsChooseAction;
 const defaultValue: ToolsChoose = {
   cursor: 'arrow',
   entity: 'none',
-  mapping: 0,
+  mapping: 'none',
 };
 
 export const useDrawToolsStore = create<DrawToolsSlice>()((set) => ({
